fix: separate state-tracked objects
